Enhanced algorithm
I found that the current program was not a very good demonstration of
thread starvation.

Hence, the program was revamped :

* The program now runs 5 threads instead of 3.
* The main thread halts for 10 minutes instead of 5, yielding more
comprehensive results.
* All but one of the threads are given the Highest priority level. The
remaining thread is given the Lowest priority.

With these changes, the output would be a more effective demonstration
of thread starvation, in which one thread suffers with the least amount
of CPU time given for it to perform its tasks.